+++ classes.h
@@ -137,10 +137,10 @@ public:
 
 class CreditCard : public Card {
 public:
-	CreditCard(const size_t id, const std::string& cardNumber,
-		const std::string& cvv, const size_t ownerId, const double currentBalance,
-		const size_t expireDate, const float transactionCommission,
-		const float witdrawCommission, const size_t creditLimit, const std::string& pin)
+	CreditCard(const uint32_t id, const std::string& cardNumber,
+               const std::string& cvv, const uint32_t ownerId, const double currentBalance,
+               const uint32_t expireDate, const float transactionCommission,
+               const float witdrawCommission, const uint32_t creditLimit, const std::string& pin)
 		: Card(id, cardNumber, cvv, ownerId, currentBalance, expireDate, CardType::Credit, transactionCommission, witdrawCommission, pin),
 		_creditLimit(creditLimit)
 	{}
